refactor: simplify sub-API class access
