+++ src/CloudAPI/index.ts
@@ -7,19 +7,40 @@
  * @see    https://greatdetail.com
  */
 import AbstractAPI, { AbstractAPIParameters } from "../API/AbstractAPI.js";
-import CloudAPIInvalidParameterError from "./CloudAPIInvalidParameterError.js";
-import CloudAPIMedia, {
-  CloudAPIMediaParameters,
-} from "./CloudAPIMedia/index.js";
+import CloudAPIMedia from "./CloudAPIMedia/index.js";
 import CloudAPIMessage, {
-  CloudAPIMessageParameters as BaseCloudAPIMessageParameters,
+  CloudAPIMessageParameters,
 } from "./CloudAPIMessage/index.js";
 import CloudAPIWebhook from "./CloudAPIWebhook/index.js";
+import { randomBytes } from "node:crypto";
 
 export interface WhatsAppAPIParameters extends AbstractAPIParameters {}
 
-export interface CloudAPIMessageParameters
-  extends Partial<BaseCloudAPIMessageParameters> {}
+export interface WhatsAppAPICreateVerifyTokenParameters {
+  /**
+   * The length of the verify token.
+   *
+   * @since 5.1.0
+   * @default 16
+   */
+  length?: number;
+
+  /**
+   * The encoding of the verify token.
+   *
+   * @since 5.1.0
+   * @default hex
+   */
+  encoding?: BufferEncoding;
+
+  /**
+   * Random Bytes Generation.
+   *
+   * @since 5.1.0
+   * @default crypto.randomBytes
+   */
+  random?: (length: number) => Buffer;
+}
 
 /**
  * WhatsApp Cloud API SDK.
@@ -32,6 +53,20 @@ export interface CloudAPIMessageParameters
  */
 export default class CloudAPI extends AbstractAPI {
   /**
+   * Default Verify Token Length.
+   *
+   * @since 5.6.0
+   */
+  public static DEFAULT_VERIFY_TOKEN_LENGTH = 16;
+
+  /**
+   * Default Verify Token Encoding.
+   *
+   * @since 5.6.0
+   */
+  public static DEFAULT_VERIFY_TOKEN_ENCODING: BufferEncoding = "hex";
+
+  /**
    * Webhook API.
    * Receive and handle messages from WhatsApp via WebHook.
    *
@@ -39,9 +74,22 @@ export default class CloudAPI extends AbstractAPI {
    */
   public webhook: CloudAPIWebhook;
 
+  /**
+   * Media API.
+   *
+   * @example
+   * // Upload Media
+   * const upload = sdk.media({ businessID: "123456" })
+   *   .upload(file);
+   * const uploadReceipt = await upload.send();
+   * console.log(uploadReceipt);
+   */
+  public media;
+
   constructor(parameters: WhatsAppAPIParameters = {}) {
     super(parameters);
     this.webhook = new CloudAPIWebhook(parameters);
+    this.media = new CloudAPIMedia(parameters);
   }
 
   /**
@@ -54,44 +102,23 @@ export default class CloudAPI extends AbstractAPI {
    * const sendReceipt = await message.send();
    * console.log(sendReceipt);
    */
-  public message({
-    businessID: overrideBusinessID,
-    logger,
-    ...parameters
-  }: CloudAPIMessageParameters = {}) {
-    const businessID = overrideBusinessID ?? this._businessID;
-    if (!businessID) {
-      throw new CloudAPIInvalidParameterError("Business ID is required");
-    }
-
-    return new CloudAPIMessage({
-      logger: logger ?? this._logger,
-      businessID,
-      ...parameters,
-    });
+  public message(parameters: CloudAPIMessageParameters) {
+    return new CloudAPIMessage({ ...parameters });
   }
 
   /**
-   * Media API.
+   * Create a new Verify Token.
+   * This is a random string that is used to verify that the request is coming
+   * from WhatsApp. This method **only** creates the value, it's usage is up to
+   * the implementer.
    *
-   * @example
-   * // Upload Media
-   * const upload = sdk.media({ businessID: "123456" })
-   *   .upload(file);
-   * const uploadReceipt = await upload.send();
-   * console.log(uploadReceipt);
+   * @since 5.6.0
    */
-  public media({
-    businessID: overrideBusinessID,
-    logger,
-    ...parameters
-  }: CloudAPIMediaParameters = {}) {
-    const businessID = overrideBusinessID ?? this._businessID;
-
-    return new CloudAPIMedia({
-      logger: logger ?? this._logger,
-      businessID,
-      ...parameters,
-    });
+  public static createVerifyToken({
+    length = this.DEFAULT_VERIFY_TOKEN_LENGTH,
+    encoding = this.DEFAULT_VERIFY_TOKEN_ENCODING,
+    random = randomBytes,
+  }: WhatsAppAPICreateVerifyTokenParameters = {}): string {
+    return random(length).toString(encoding);
   }
 }
